Fix C# script compilation.
This probably isn't the best way to actually fix this, but it at least demonstrates what the problem was. If scripts are meant to link with the main framework DLL, it should probably forward stuff from the scripting DLL, and if they're meant to use the scripting DLL, the default value for the setting should be the scripting DLL, not the framework one.